requirements.txt
graia-application-mirai 0.18.4
graia-broadcast         0.8.11